add validations related to Projects

+++ src/core/useCases/Project/DeleteProject.ts
@@ -8,6 +8,12 @@ export class DeleteProject {
 	}
 
 	async exec(id: number) {
+		const project = await this.repository.findProjectById(id)
+
+		if (!project) {
+			throw new Error('This project not exists')
+		}
+		
 		await this.repository.deleteProject(id)
 	}
 }
